Add files via upload
Implement a problem of minimum work to be done per day to 
finish given tasks within D days problem. 
Statement: Given an array task[] of size N denoting amount of 
work to be done for each task, the problem is to find the minimum 
amount of work to be done on each day so that all the tasks can be 
completed in at most D days. Note: On one day work can be done 
for only one task. 
Input: task[] = [3, 4, 7, 15],  D = 10                     Output: 4 
Input: task[] = [30, 20, 22, 4, 21], D = 6              Output: 22